requirements.txt
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
zenpy==2.0.12
